+++ backend/reviewduck/src/test/java/com/reviewduck/common/service/ServiceTest.java
@@ -1,4 +1,4 @@
-package com.reviewduck;
+package com.reviewduck.common.service;
 
 import javax.transaction.Transactional;
