+++ backend/reviewduck/src/test/java/com/reviewduck/member/acceptance/MemberAcceptanceTest.java
@@ -8,9 +8,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import com.reviewduck.acceptance.AcceptanceTest;
+import com.reviewduck.common.acceptance.AcceptanceTest;
 import com.reviewduck.member.dto.request.MemberUpdateNicknameRequest;
-import com.reviewduck.member.dto.response.MemberDto;
 import com.reviewduck.member.dto.response.MemberResponse;
 
 public class MemberAcceptanceTest extends AcceptanceTest {
